Move checked trade mutations into TradeItemContainer
